+++ changelog.md
@@ -13,6 +13,7 @@
 - Games show as columns on wide screens to save space
 - Removed confirmation dialogue when adding unregistered players to stages
 - New card in Faction settings to view the faction in the Forge if linked
+- Tournament Stages can now enable rounds independent of the Tournament
 
 ### Bug Fixes
 - Expansion image bug where the image could be unintentionally deleted